backend/requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12

# Async Tasks
celery[redis]==5.4.0
redis==5.1.1

# Vector DB
qdrant-client==1.11.3

# Document Extraction
PyMuPDF==1.24.10
openpyxl==3.1.5

# Embeddings
sentence-transformers==3.1.1

# Keyword Extraction (reemplaza al LLM generativo)
keybert==0.8.5

# LLM Providers (optional, loaded dynamically based on LLM_PROVIDER env var)
openai>=1.0.0
google-generativeai>=0.8.0
anthropic>=0.30.0

# Config
pydantic-settings==2.5.2
python-dotenv==1.0.1
pytesseract
pdf2image

# Auth
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
